LayoutTests/http/tests/security/xssAuditor/full-block-object-tag-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

There should be no content in the iframe below:



--------
Frame: 'frame'
--------
